pyproject.toml
[tool.poetry]
name = "schedry-bot"
version = "0.1.0"
description = ""
authors = ["jieggii <jieggii.contact@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = {extras = ["cli"], version = "^0.19.1"}
aiohttp = "^3.7.4"
betterconf = "^2.6.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.8"}
pydantic = "^1.8.2"
vkwave = {git = "https://github.com/fscdev/vkwave.git", rev = "4c6945b"}
uvloop = "^0.16.0"

[tool.black]
line-length = 120
target-version = ['py39']

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
